PW_FT_detection/requirements.txt
PytorchWildlife
ultralytics
hydra
wget